🚚 move:잡담게시글 파일 이동
