+++ src/main/java/com/example/backend/community/repository/CommunityBoardRepository.java
@@ -1,5 +1,6 @@
-package com.example.backend.community;
+package com.example.backend.community.repository;
 
+import com.example.backend.community.domain.CommunityBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
