+++ src/main/java/com/example/backend/security/WebSecurityConfig.java
@@ -13,6 +13,7 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
             "/","/**/*.png","/**/*.jpg","/**/*.js","/**/*.css","/**/*.html","/**/*.gif","/**/*.svg"
             ,"/signup","/signup/**","/signin","/signin/**"//소셜 로그인 관련한 부분도 추가
             ,"/test","/test-db","/test-db/**" //s3, rds 연동 테스트때 사용한 url 들 일단 접근 허가
+            ,"/boards/community","/boards/community/**" //로그인 구현 전 잡담게시판 구현 위해 임시 허가
     };
 
     @Override
